refactor(storageListeners): Added to the Storage Management custom event's types/names as optionals parameter.
Added an optional parameter for the `storageManagement.ts` functions: `event`, which can be just a single string or an array of strings. It can be used/understand as a "better connection" of events that can happen to the localStorage and components and functions that depend on these events.

For example: now components like the `ProgressBar.vue` just need to listen to changes on tasks and sub-tasks states changes and not listen to every update on the localStorage.